rework trigger
Renamed NetworkTrigger to MessageTrigger and add the new ConectionTrigger.
The MessageTigger itself is still untouched. The renaming was only for better
separation with the new tigger. The ConnectionTrigger is for incoming
connnections at the server. Similar to the MessageTigger, it is called for
every incoming connection. That way, other components can decide, what
should be done with the new created socket.
Additionally servers now have only one message- and one connection-trigger.
Not a list anymore.
Also updated tests to use the new trigger-class.